fix tests and imports
