--- mcp_servers/database_service/__init__.py
@@ -1,1 +1,0 @@
-# Database Service MCP Server
