+++ src/storage.py
@@ -13,18 +13,21 @@ class Base(DeclarativeBase):
 
 
 def build_database_url() -> str:
-    if envs.STORAGE_DB == "mysql":
+    if envs.STORAGE_DB.startswith("sqlite"):
         if not envs.DEBUG_MODE:
             logger.warning(
-                "The `mysql` is used as storage db! Use it ONLY for development"
+                "The `sqlite` is used as storage db! Use it ONLY for development"
             )
-        return "sqlite:///./dev.db"
+
+        if envs.STORAGE_DB == "sqlite-memory":
+            return "sqlite:///:memory:"
+        else:
+            return "sqlite:///./dev.db"
     elif envs.STORAGE_DB == "postgres":
         logger.info("The `postgres` is used as storage db")
         return f"postgresql+psycopg2://{envs.PG_USER}:{envs.PG_PASSWORD}@{envs.PG_HOST}:{envs.PG_PORT}/telegram_bot"
     else:
         raise ValueError("The `STORAGE_DB` env variable is not correct")
-
 
 
 def get_engine_and_sessionmaker() -> Tuple[object, sessionmaker]:
@@ -55,6 +58,6 @@ def get_db_session(SessionLocal: sessionmaker) -> Callable:
     return _get_db
 
 
-
 engine, SessionLocal = get_engine_and_sessionmaker()
+init_db(engine)
 get_db = get_db_session(SessionLocal)
