+++ test/user_group/test_group.py
@@ -1,20 +1,6 @@
 import pytest
 
-from storage import SessionLocal, init_db, engine, Base
 from user_group_db.models import User, Group
-
-
-@pytest.fixture
-def session():
-    with SessionLocal() as session:
-        yield session
-
-    # need to drop all tables after each test
-    # and re-create them
-    # otherwise data from previous test will be
-    # present in the next test
-    Base.metadata.drop_all(bind=engine)
-    init_db(engine)
 
 
 def test_create_group(session):
